defi-ai-watchown/requirements.txt
web3==6.15.1
groq
python-dotenv==1.0.1
requests==2.31.0
